loadorder refac
